backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# AI/ML Libraries
torch>=2.2.0
transformers>=4.36.2
scikit-learn>=1.3.2
numpy>=1.24.4
pandas>=2.1.4
scipy>=1.11.4

# NLP Libraries
spacy>=3.7.2
nltk>=3.8.1

# Time-series forecasting
prophet>=1.1.4
tensorflow>=2.15.0

# API and HTTP
requests>=2.31.0
httpx>=0.25.2
pydantic>=2.5.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Environment and config
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Data processing
plotly>=5.17.0
seaborn>=0.13.0
matplotlib>=3.8.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Development
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0

# Deployment
gunicorn>=21.2.0